kept SDL1 as it was
