+++ CMake/platforms/linux.cmake
@@ -1,1 +1,2 @@
-set(FULL_RANGE_CONTROLLER ON)
+# for future use
+# set(OPTIONS_SET_GAME_CONTROLLER ON)
