+++ CMake/platforms/switch.cmake
@@ -16,3 +16,5 @@ set(JOY_BUTTON_DPAD_DOWN 19)
 
 list(APPEND DEVILUTIONX_PLATFORM_SUBDIRECTORIES platform/switch)
 list(APPEND DEVILUTIONX_PLATFORM_LINK_LIBRARIES libdevilutionx_switch)
+
+set(SWAP_CONFIRM_CANCEL_BUTTONS ON)
