+++ CMake/platforms/windows.cmake
@@ -21,4 +21,5 @@ if(MINGW_CROSS)
   list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}/mingw")
 endif()
 
-set(FULL_RANGE_CONTROLLER ON)
+# for future use
+# set(OPTIONS_SET_GAME_CONTROLLER ON)
